Convert examples to scala-cli tutorials
