+++ examples/6-image.sc
