+++ examples/assets/scala.bmp
